backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-jazzmin==3.0.1
djangorestframework==3.16.1
pillow==12.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
